Extensions update.

+++ setup.sh
@@ -1,3 +1,6 @@
 #!/bin/bash
-stack build --copy-compiler-tool intero phoityne-vscode hindent hlint
-brew install rabbitmq libpq postgres
+if [ ! -f ~/.vest_haskell ]; then
+  stack build --copy-compiler-tool intero phoityne-vscode hindent hlint
+  brew install rabbitmq libpq postgres
+  touch ~/.vest_haskell
+fi
